+++ packages/client/src/App.tsx
@@ -1,9 +1,11 @@
 import { observer } from 'mobx-react-lite';
-import React from 'react';
+import React, { useEffect } from 'react';
 
 import { UIWrapper } from '$components/ui';
+import { mainStore } from '$stores/MainStore';
 import { setupStore } from '$stores/SetupStore';
 import { uiStore } from '$stores/UIStore';
+import { withoutTransitions } from '$utils/css';
 import './index.scss';
 import { Home } from './views/home';
 import { ModelManager } from './views/models';
@@ -13,6 +15,12 @@ import { Setup } from './views/setup';
 import { Main } from './views/system/Main';
 
 const View: React.FC = observer(() => {
+  useEffect(() => {
+    withoutTransitions(() => {
+      document.documentElement.setAttribute('data-theme', mainStore.theme);
+    });
+  }, [mainStore.theme]);
+
   if (setupStore?.status !== 'done') {
     return <Setup />;
   }