+++ packages/client/src/components/toggle/index.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { BsCheck } from 'react-icons/bs';
 
 import styles from './index.module.scss';
 
@@ -27,14 +28,7 @@ export const Toggle: React.FC<ToggleProps> = ({
       <span className={styles.track}>
         <span className={styles.indicator}>
           <span className={styles.checkmark}>
-            <svg
-              viewBox="0 0 24 24"
-              id="ghq-svg-check"
-              role="presentation"
-              aria-hidden="true"
-            >
-              <path d="M9.86 18a1 1 0 01-.73-.32l-4.86-5.17a1.001 1.001 0 011.46-1.37l4.12 4.39 8.41-9.2a1 1 0 111.48 1.34l-9.14 10a1 1 0 01-.73.33h-.01z"></path>
-            </svg>
+            <BsCheck />
           </span>
         </span>
       </span>